Section 08: Arrays, Java inbuilt Lists, Auto boxing and Unboxing - Linked list topic covered
Solving following coding exercises:
- Coding Exercise 46: Playlist

Renaming coding exercises folders
