+++ src/coding_exercises/coding_exercise_02_MegaBytesConverter/MegaBytesConverter.java
@@ -1,0 +1,21 @@
+package coding_exercises.coding_exercise_02_MegaBytesConverter;
+
+public class MegaBytesConverter {
+
+    public static void main(String[] args) {
+        printMegaBytesAndKiloBytes(2500);
+        printMegaBytesAndKiloBytes(-1024);
+        printMegaBytesAndKiloBytes(5000);
+    }
+
+    public static void printMegaBytesAndKiloBytes(int kiloBytes) {
+        int megaBytes = kiloBytes / 1024;
+        int remainingKiloBytes = kiloBytes % 1024;
+
+        if (kiloBytes < 0) {
+            System.out.println("Invalid Value");
+        } else {
+            System.out.printf("%d KB = %d MB and %d KB%n", kiloBytes, megaBytes, remainingKiloBytes);
+        }
+    }
+}
