--- src/coding_exercises/coding_exercise_21/AllFactors.java
@@ -1,20 +1,0 @@
-package coding_exercises.coding_exercise_21;
-
-public class AllFactors {
-
-    public static void main(String[] args) {
-        printFactors(32);
-    }
-
-    public static void printFactors(int number) {
-        if (number < 1) {
-            System.out.println("Invalid Value");
-        } else {
-            for (int i = 1; i <= number; i++) {
-                if (number % i == 0) {
-                    System.out.println(i);
-                }
-            }
-        }
-    }
-}
