+++ src/coding_exercises/coding_exercise_46_Playlist/Song.java
@@ -1,0 +1,21 @@
+package coding_exercises.coding_exercise_46_Playlist;
+
+public class Song {
+
+    private String title;
+    private double duration;
+
+    public Song(String title, double duration) {
+        this.title = title;
+        this.duration = duration;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    @Override
+    public String toString() {
+        return this.title + ": " + this.duration;
+    }
+}
