+++ src/s08_arrays_inbuiltlists_autoboxing_unboxing/linkedlist/Customer.java
@@ -1,0 +1,27 @@
+package s08_arrays_inbuiltlists_autoboxing_unboxing.linkedlist;
+
+public class Customer {
+    private String name;
+    private double balance;
+
+    public Customer(String name, double balance) {
+        this.name = name;
+        this.balance = balance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getBalance() {
+        return balance;
+    }
+
+    public void setBalance(double balance) {
+        this.balance = balance;
+    }
+}
