+++ src/s08_arrays_inbuiltlists_autoboxing_unboxing/linkedlist/Demo.java
@@ -1,0 +1,31 @@
+package s08_arrays_inbuiltlists_autoboxing_unboxing.linkedlist;
+
+import java.util.Iterator;
+import java.util.LinkedList;
+
+public class Demo {
+
+    public static void main(String[] args) {
+        LinkedList<String> placesToVisit = new LinkedList<>();
+        placesToVisit.add("Sydney");
+        placesToVisit.add("Melbourne");
+        placesToVisit.add("Brisbane");
+        placesToVisit.add("Perth");
+        placesToVisit.add("Canberra");
+        placesToVisit.add("Adelaide");
+        placesToVisit.add("Darwin");
+        
+        printList(placesToVisit);
+
+        placesToVisit.add(1, "Alice Spring");
+        printList(placesToVisit);
+    }
+
+    private static void printList(LinkedList<String> list) {
+        Iterator<String> i = list.iterator();
+        while (i.hasNext()) {
+            System.out.println("Now visiting " + i.next());
+        }
+        System.out.println("=============================");
+    }
+}
